+++ .gitignore
@@ -53,8 +53,9 @@ typings/
 # Output of 'npm pack'
 *.tgz
 
-# Yarn Integrity file
+# Yarn files
 .yarn-integrity
+yarn.lock
 
 # dotenv environment variables file
 .env
