+++ jupyterlab_gcpscheduler/package.json
@@ -51,7 +51,7 @@
     "@phosphor/signaling": "^1.3.0",
     "csstips": "^1.2.0",
     "formik": "^1.5.8",
-    "gcp-jupyterlab-shared": "^1.0.0",
+    "gcp-jupyterlab-shared": "^1.0.1",
     "typestyle": "^2.1.0"
   },
   "devDependencies": {
