+++ jupyterlab_gcpscheduler/src/service/gcp.spec.ts
@@ -19,7 +19,8 @@ import { ServerConnection } from '@jupyterlab/services';
 import { IDocumentManager } from '@jupyterlab/docmanager';
 import { CLOUD_FUNCTION_REGION, IMPORT_DIRECTORY } from '../data';
 import { GcpService, RunNotebookRequest } from './gcp';
-import { TEST_PROJECT, asApiResponse } from '../test_helpers';
+import { TEST_PROJECT } from '../test_helpers';
+import { asApiResponse } from 'gcp-jupyterlab-shared';
 import { ProjectStateService } from './project_state';
 import { Contents } from '@jupyterlab/services';
 import { INotebookModel, Notebook } from '@jupyterlab/notebook';
