+++ jupyterlab_gcpscheduler/src/test_helpers.ts
@@ -15,7 +15,6 @@
  */
 
 /** Utility functions and helpers for tests. */
-import { ApiResponse } from './service/transport';
 import { ProjectState } from './service/project_state';
 import { ReactWrapper, ShallowWrapper } from 'enzyme';
 import { AiPlatformJob } from './service/gcp';
@@ -99,19 +98,6 @@ export function triggeredResolver(
     done = true;
   };
   return { resolve, promise };
-}
-
-/** Returns a Promise that resolves a JSON response akin to the fetch API */
-export function asFetchResponse(result: any, ok = true): Promise<Response> {
-  return Promise.resolve({
-    ok,
-    json: () => Promise.resolve(result),
-  } as Response);
-}
-
-/** Wraps an object in the shape of a transport API response */
-export function asApiResponse<T>(body: T): ApiResponse<T> {
-  return { result: body };
 }
 
 // Simulates a form input change
