+++ shared/client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "gcp-jupyterlab-shared",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "Shared libraries for JupyterLab extensions",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
@@ -31,6 +31,7 @@
   "dependencies": {
     "@material-ui/core": "^3.9.4",
     "@material-ui/icons": "^3.0.2",
+    "@jupyterlab/services": "^4.2.2",
     "csstips": "^1.2.0",
     "typestyle": "^2.1.0",
     "react": "~16.8.4",
@@ -40,6 +41,7 @@
     "@types/jest": "^24.9.1",
     "@types/react": "~16.8.18",
     "@types/react-dom": "^16.8.4",
+    "@types/gapi": "0.0.39",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
     "enzyme-to-json": "^3.4.4",
